+++ packages/components-angular/src/app/app-routing.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { RouterModule, type Route } from '@angular/router';
+import { type Route, RouterModule } from '@angular/router';
 import * as fromExamples from './examples';
 import * as fromPages from './pages';
 import * as fromStyles from './styles';
@@ -94,6 +94,16 @@ export const routes: ExtendedRoute[] = [
     name: 'Accordion Example',
     path: 'accordion-example',
     component: fromExamples.AccordionExampleComponent,
+  },
+  {
+    name: 'AG Grid Example',
+    path: 'ag-grid-example',
+    component: fromExamples.AgGridExampleComponent,
+  },
+  {
+    name: 'AG Grid Example Storefront',
+    path: 'ag-grid-example-storefront',
+    component: fromExamples.AgGridExampleStorefrontComponent,
   },
   {
     name: 'Banner Example',
@@ -274,6 +284,16 @@ export const routes: ExtendedRoute[] = [
     name: 'Textarea Example Controlled',
     path: 'textarea-controlled-example',
     component: fromExamples.TextareaExampleControlledComponent,
+  },
+  {
+    name: 'Checkbox Example',
+    path: 'checkbox-example',
+    component: fromExamples.CheckboxExampleComponent,
+  },
+  {
+    name: 'Checkbox Example Controlled',
+    path: 'checkbox-controlled-example',
+    component: fromExamples.CheckboxExampleControlledComponent,
   },
   {
     name: 'Text Field Wrapper Example IMask',