+++ packages/components-angular/tests/e2e/specs/examples.e2e.ts
@@ -23,6 +23,8 @@ const exampleUrls = exampleRoutes.map((item) => item.path);
 
 for (const exampleUrl of exampleUrls) {
   test(`should work without error or warning for ${exampleUrl}`, async ({ page }) => {
+    // Skip AG Grid pages since they will show licensing errors
+    test.skip(['ag-grid-example', 'ag-grid-example-storefront'].includes(exampleUrl));
     await goto(page, exampleUrl);
     expect(getConsoleErrorsAmount()).toBe(0);
     expect(getConsoleWarningsAmount()).toBe(0);