+++ packages/components/src/components.d.ts
@@ -14,6 +14,7 @@ import { ButtonPureAlignLabel, ButtonPureAriaAttribute, ButtonPureIcon, ButtonPu
 import { ButtonTileAlign, ButtonTileAriaAttribute, ButtonTileAspectRatio, ButtonTileBackground, ButtonTileIcon, ButtonTileSize, ButtonTileType, ButtonTileWeight } from "./components/button-tile/button-tile-utils";
 import { CanvasSidebarEndIcon, CanvasSidebarStartIcon } from "./components/canvas/canvas-utils";
 import { CarouselAlignHeader, CarouselAriaAttribute, CarouselGradientColor, CarouselHeadingSize, CarouselInternationalization, CarouselUpdateEventDetail, CarouselWidth } from "./components/carousel/carousel-utils";
+import { CheckboxBlurEventDetail, CheckboxState, CheckboxUpdateEventDetail } from "./components/checkbox/checkbox-utils";
 import { CheckboxWrapperState } from "./components/checkbox-wrapper/checkbox-wrapper-utils";
 import { ContentWrapperBackgroundColor, ContentWrapperWidth } from "./components/content-wrapper/content-wrapper-utils";
 import { CrestAriaAttribute, CrestTarget } from "./components/crest/crest-utils";
@@ -77,6 +78,7 @@ export { ButtonPureAlignLabel, ButtonPureAriaAttribute, ButtonPureIcon, ButtonPu
 export { ButtonTileAlign, ButtonTileAriaAttribute, ButtonTileAspectRatio, ButtonTileBackground, ButtonTileIcon, ButtonTileSize, ButtonTileType, ButtonTileWeight } from "./components/button-tile/button-tile-utils";
 export { CanvasSidebarEndIcon, CanvasSidebarStartIcon } from "./components/canvas/canvas-utils";
 export { CarouselAlignHeader, CarouselAriaAttribute, CarouselGradientColor, CarouselHeadingSize, CarouselInternationalization, CarouselUpdateEventDetail, CarouselWidth } from "./components/carousel/carousel-utils";
+export { CheckboxBlurEventDetail, CheckboxState, CheckboxUpdateEventDetail } from "./components/checkbox/checkbox-utils";
 export { CheckboxWrapperState } from "./components/checkbox-wrapper/checkbox-wrapper-utils";
 export { ContentWrapperBackgroundColor, ContentWrapperWidth } from "./components/content-wrapper/content-wrapper-utils";
 export { CrestAriaAttribute, CrestTarget } from "./components/crest/crest-utils";
@@ -498,6 +500,64 @@ export namespace Components {
           * @deprecated since v3.0.0, will be removed with next major release
          */
         "wrapContent"?: boolean;
+    }
+    interface PCheckbox {
+        /**
+          * Reflects the checkbox current checked state and allows setting the initial checked state.
+         */
+        "checked"?: boolean;
+        /**
+          * Displays as compact version.
+         */
+        "compact"?: boolean;
+        /**
+          * Marks the checkbox as disabled.
+         */
+        "disabled"?: boolean;
+        /**
+          * The id of a form element the checkbox should be associated with.
+         */
+        "form"?: string;
+        /**
+          * Show or hide label. For better accessibility it's recommended to show the label.
+         */
+        "hideLabel"?: BreakpointCustomizable<boolean>;
+        /**
+          * Marks the checkbox as indeterminate.
+         */
+        "indeterminate"?: boolean;
+        /**
+          * The label text.
+         */
+        "label"?: string;
+        /**
+          * @experimental Disables the checkbox and shows a loading indicator.
+         */
+        "loading"?: boolean;
+        /**
+          * The message styled depending on validation state.
+         */
+        "message"?: string;
+        /**
+          * The name of the checkbox.
+         */
+        "name"?: string;
+        /**
+          * Marks the checkbox as required.
+         */
+        "required"?: boolean;
+        /**
+          * The validation state.
+         */
+        "state"?: CheckboxState;
+        /**
+          * Adapts the color depending on the theme.
+         */
+        "theme"?: Theme;
+        /**
+          * The checkbox value. When a form is submitted, only a checkbox which is currently checked is included in the submission.
+         */
+        "value"?: string;
     }
     interface PCheckboxWrapper {
         /**
@@ -2225,6 +2285,10 @@ export interface PCarouselCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPCarouselElement;
 }
+export interface PCheckboxCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLPCheckboxElement;
+}
 export interface PFlyoutCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPFlyoutElement;
@@ -2392,6 +2456,24 @@ declare global {
     var HTMLPCarouselElement: {
         prototype: HTMLPCarouselElement;
         new (): HTMLPCarouselElement;
+    };
+    interface HTMLPCheckboxElementEventMap {
+        "update": CheckboxUpdateEventDetail;
+        "blur": CheckboxBlurEventDetail;
+    }
+    interface HTMLPCheckboxElement extends Components.PCheckbox, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLPCheckboxElementEventMap>(type: K, listener: (this: HTMLPCheckboxElement, ev: PCheckboxCustomEvent<HTMLPCheckboxElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLPCheckboxElementEventMap>(type: K, listener: (this: HTMLPCheckboxElement, ev: PCheckboxCustomEvent<HTMLPCheckboxElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLPCheckboxElement: {
+        prototype: HTMLPCheckboxElement;
+        new (): HTMLPCheckboxElement;
     };
     interface HTMLPCheckboxWrapperElement extends Components.PCheckboxWrapper, HTMLStencilElement {
     }
@@ -3081,6 +3163,7 @@ declare global {
         "p-button-tile": HTMLPButtonTileElement;
         "p-canvas": HTMLPCanvasElement;
         "p-carousel": HTMLPCarouselElement;
+        "p-checkbox": HTMLPCheckboxElement;
         "p-checkbox-wrapper": HTMLPCheckboxWrapperElement;
         "p-content-wrapper": HTMLPContentWrapperElement;
         "p-crest": HTMLPCrestElement;
@@ -3536,6 +3619,72 @@ declare namespace LocalJSX {
           * @deprecated since v3.0.0, will be removed with next major release
          */
         "wrapContent"?: boolean;
+    }
+    interface PCheckbox {
+        /**
+          * Reflects the checkbox current checked state and allows setting the initial checked state.
+         */
+        "checked"?: boolean;
+        /**
+          * Displays as compact version.
+         */
+        "compact"?: boolean;
+        /**
+          * Marks the checkbox as disabled.
+         */
+        "disabled"?: boolean;
+        /**
+          * The id of a form element the checkbox should be associated with.
+         */
+        "form"?: string;
+        /**
+          * Show or hide label. For better accessibility it's recommended to show the label.
+         */
+        "hideLabel"?: BreakpointCustomizable<boolean>;
+        /**
+          * Marks the checkbox as indeterminate.
+         */
+        "indeterminate"?: boolean;
+        /**
+          * The label text.
+         */
+        "label"?: string;
+        /**
+          * @experimental Disables the checkbox and shows a loading indicator.
+         */
+        "loading"?: boolean;
+        /**
+          * The message styled depending on validation state.
+         */
+        "message"?: string;
+        /**
+          * The name of the checkbox.
+         */
+        "name"?: string;
+        /**
+          * Emitted when the checkbox has lost focus.
+         */
+        "onBlur"?: (event: PCheckboxCustomEvent<CheckboxBlurEventDetail>) => void;
+        /**
+          * Emitted when checkbox checked property is changed.
+         */
+        "onUpdate"?: (event: PCheckboxCustomEvent<CheckboxUpdateEventDetail>) => void;
+        /**
+          * Marks the checkbox as required.
+         */
+        "required"?: boolean;
+        /**
+          * The validation state.
+         */
+        "state"?: CheckboxState;
+        /**
+          * Adapts the color depending on the theme.
+         */
+        "theme"?: Theme;
+        /**
+          * The checkbox value. When a form is submitted, only a checkbox which is currently checked is included in the submission.
+         */
+        "value"?: string;
     }
     interface PCheckboxWrapper {
         /**
@@ -5394,6 +5543,7 @@ declare namespace LocalJSX {
         "p-button-tile": PButtonTile;
         "p-canvas": PCanvas;
         "p-carousel": PCarousel;
+        "p-checkbox": PCheckbox;
         "p-checkbox-wrapper": PCheckboxWrapper;
         "p-content-wrapper": PContentWrapper;
         "p-crest": PCrest;
@@ -5486,6 +5636,7 @@ declare module "@stencil/core" {
              * @controlled { "props": ["activeSlideIndex"], "event": "update", "isInternallyMutated": true }
              */
             "p-carousel": LocalJSX.PCarousel & JSXBase.HTMLAttributes<HTMLPCarouselElement>;
+            "p-checkbox": LocalJSX.PCheckbox & JSXBase.HTMLAttributes<HTMLPCheckboxElement>;
             "p-checkbox-wrapper": LocalJSX.PCheckboxWrapper & JSXBase.HTMLAttributes<HTMLPCheckboxWrapperElement>;
             /**
              * @deprecated since v3.0.0, will be removed with next major release. Use native CSS Grid instead.